% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime_functions.R
\name{append_wibrs}
\alias{append_wibrs}
\title{append_wibrs}
\usage{
append_wibrs(old, new)
}
\arguments{
\item{old}{A data frame of WIBRS data.}

\item{new}{A data frame of WIBRS data.}
}
\value{
A data frame.
}
\description{
append_wibrs appends new, raw WIBRS data to the historical, raw WIBRS dataset.
It assumes that the new data frame came from an Excel file downloaded from
the City of Milwaukee's COMPASS server (see get_wibrs) and therefore trims
the last three rows of the new data frame.
}
\examples{
data(raw_wibrs_2005thru2015)
wibrs2015 <- xlsx::read.xlsx2("~/Dropbox/Analytics-Consulting/Non-client_projects/MKE data/crime/2015_WIBR_Radius99K_accessed041316.xls", 1) # accessed 04/13/16
# Note: the following example results in dupes
combo <- append_wibrs(raw_wibrs_2005thru2015, wibrs2015)
}

