% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{createCentroidTable}
\alias{createCentroidTable}
\title{get_mai}
\usage{
createCentroidTable(longitude, latitude, radius, max_radius_km)
}
\arguments{
\item{longitude}{Standard Longitude in range -/+180}

\item{latitude}{Standard latitude in range -/+90}

\item{radius}{Radius in kilometers of the large circle to be covered by small cricles}

\item{max_radius_km}{Radius in kilometers of the small cricles (e.g. 40km)}
}
\value{
A data frame of the (longitude,latitude) pairs for the centres of the small circles which cover the large circle
}
\description{
get_mai retrieves the Master Index file from the City of Milwaukee's server. This speeds up geocoding of large files.
}
\examples{
centre_table <- createCentroidTable(-122,37,400,40)
plot(centre_table,pch=18)

}

