% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{crime.geo}
\alias{crime.geo}
\title{crime.geo dataset}
\format{An object of class \code{data.frame} with 546464 rows and 15 columns.}
\usage{
crime.geo
}
\description{
This contains all WIBRS crime incident records for the entire City of
Milwukee, from Jan 1, 2005 through Dec 31 2015. Records were downloaded from
the COMPASS website. 2015 data were retreived on 04/13/2016. Data was then
geocoded with full_geocode(). X/Y coordinates are in the NAD27 datum. Raw
WIBRS data is in raw_wibrs_2005thru2015 dataset.
}
\examples{
data(crime.geo)
head(crime.geo)

}
\author{
Matthew Schumwinger
}
\keyword{datasets}

