% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{crimes.munged}
\alias{crimes.munged}
\title{crimes.munged dataset}
\format{An object of class \code{SpatialPointsDataFrame} with 532941 rows and 18 columns.}
\usage{
crimes.munged
}
\description{
This contains all WIBRS crime incident records for the entire City of
Milwukee, from Jan 1, 2005 through Dec 31 2015. Records were downloaded from
the COMPASS website. 2015 data were retreived on 04/13/2016. Data was then
geocoded with full_geocode(). X/Y coordinates are in the NAD27 datum.
}
\details{
The dataframe was then processed with munge_wibrs() and transformed into a
SpatialPointsDataFrame ready for analysis and visualization. Raw
WIBRS data is in raw_wibrs_2005thru2015 dataset.
}
\examples{
data(crimes.munged)
plot(crimes.munged)

}
\author{
Matthew Schumwinger
}
\keyword{datasets}

