% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_geocode.R
\name{full_geocode}
\alias{full_geocode}
\title{full_geocode}
\usage{
full_geocode(df, field, MAI)
}
\arguments{
\item{df}{A dataframe with City of Milwaukee records to geocode.}

\item{field}{(NOT IMPLEMENTED)Character string of single column name containing street address
information to geocode. (Note: you may need to concatenate prior to using
this function).}

\item{MAI}{A data frame containing the Master Address Index, already
geoprocessed against parcel centroids.}
}
\value{
A data frame.
}
\description{
full_geocode is a function that geocodes addresses in Milwaukee,
Wisconsin. Draft version geocodes crime data, first by matching against
geoprocessed MAI, then using low_level funciton to access City geocode API.
}
\examples{
\dontrun{
data("raw_wibrs_2005thru2015")
crime_subset <- raw_wibrs_2005thru2015[sample(nrow(raw_wibrs_2005thru2015), 30), ]
data("geo_mai")
geo_crime <- full_geocode(crime_subset, "LOCATION", geo_mai)
head(geo_crime)
}

}

