% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIgeocoder.R
\name{geo_process_mai}
\alias{geo_process_mai}
\title{geo_process_mai}
\usage{
geo_process_mai(mai = data(mai_04042016), parcels)
}
\arguments{
\item{mai}{A dataframe of the raw Master Address Index.}

\item{parcels}{A SpatialPolygonsDataFrame of parcelbase file.}
}
\value{
A data frame.
}
\description{
geo_process_mai geoprocesses the City of Milwaukee
\href{http://itmdapps.milwaukee.gov/gis/mai/Documentation/mai.pdf}{Master
Address Index (MAI)}. It does this by joining the MAI to the parcelbase file
by TAXKEY field to associate x and y (lat/lon?) values with each MAI record.
}
\examples{
\dontrun{
geo_mai <- geo_process_mai()
head(geo_mai)
}

}

