% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIgeocoder.R
\name{geoprocess_mai}
\alias{geoprocess_mai}
\title{geoprocess_mai}
\usage{
geoprocess_mai(mai, parcels)
}
\arguments{
\item{mai}{A dataframe of the raw Master Address Index.}

\item{parcels}{A SpatialPolygonsDataFrame of parcelbase file.}
}
\value{
A data frame.
}
\description{
geoprocess_mai geoprocesses the City of Milwaukee
\href{http://itmdapps.milwaukee.gov/gis/mai/Documentation/mai.pdf}{Master
Address Index (MAI)}. It does this by joining the MAI to the parcelbase file
by TAXKEY field to associate lat/lon values with each MAI record.
}
\examples{
\dontrun{
data(mai_04042016)
data(parcels)
geo_mai <- geoprocess_mai(mai_04042016, parcels)
head(geo_mai)
}

}

