% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_bids}
\alias{get_bids}
\title{get_bids}
\usage{
get_bids()
}
\value{
A large SpatialPolygonsDataFrame.
}
\description{
get_bids is a function that retreives the City of Milwaukee's Business
Improvement District (BID) kml file from the City's \href{http://maps.milwaukee.gov/ArcGIS/rest/services/planning/Special_Districts/MapServer/3/query}{GIS Services REST API}.
}
\examples{
\dontrun{
bids <- get_bids()
sp::plot(bids)
head(bids@data)
}

}

