% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIgeocoder.R
\name{get_mai}
\alias{get_mai}
\title{get_mai}
\usage{
get_mai(url = mai_url)
}
\arguments{
\item{url}{Endpoint URL. Default = \url{http://itmdapps.milwaukee.gov/xmldata/Get_mai_xml}}
}
\value{
An object of (S3) class XMLDocument. This has two fields named doc
and dtd and are of class DTDList and XMLDocumentContent respectively.
}
\description{
get_mai is a funtion that retreives the City of Milwaukee
\href{http://itmdapps.milwaukee.gov/gis/mai/Documentation/mai.pdf}{Master
Address Index (MAI)}. This retrieves the XML verion, which is approximately
5.5MB, and may take awhile to download.
}
\examples{
\dontrun{
mai <- get_mai()
head(mai)
}

}

