% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_parcels}
\alias{get_parcels}
\title{get_parcels}
\usage{
get_parcels(url = parcel_url)
}
\arguments{
\item{url}{Endpoint URL. Default = \url{http://itmdapps.milwaukee.gov/gis/mapdata/parcelbase.zip}}
}
\value{
A large SpatialPolygonsDataFrame.
}
\description{
get_parcelbase is a function that retreives the City of Milwaukee's parcel shapefile
\href{http://city.milwaukee.gov/DownloadMapData3497.htm}{from the
city's download map data page}.
}
\details{
The City of Milwaukee parcel shapefile accommodates condominium records
(multiple tax identifiers -- "taxkeys" -- assigned to one parcel).  Multiple
records (one for each taxkey) exist for condominium parcels. As a result, a
direct join can be made between parcelbase.shp and MPROP (or other
taxkey-based dataset), and condominium data will be accurately joined.
}
\examples{
\dontrun{
parcels <- get_parcels()
summary(parcels@data)
}

}

