% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime_functions.R
\name{get_wibrs}
\alias{get_wibrs}
\title{get_wibrs}
\usage{
get_wibrs(url = wibrs_url, radius_ft = 99000, center = cityhall)
}
\arguments{
\item{url}{Target url.}

\item{radius_ft}{Search radius (in feet). Default is 99000 to cover entire city.}

\item{center}{Centroid address of radius. Default is address of City Hall.}
}
\value{
A data frame.
}
\description{
get_wibrs retrieves WIBRS data from the City of Milwaukee's server.
}
\examples{
\dontrun{
raw_geo <- geocode(head(raw), "LOCATION")
raw_geo
}
}

