% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hoods}
\alias{hoods}
\title{hoods dataset}
\format{An object of class \code{SpatialPolygonsDataFrame} with 190 rows and 6 columns.}
\usage{
hoods
}
\description{
This contains the City of Milwaukee neighborhood boundary shapefile
downloaded on Feb 26, 2015 and processed to a SpatialPolygonsDataFrame.
Projected using State Plane South NAD 27 (ESPG 32054).
}
\examples{
data(hoods)
sp::plot(hoods)

}
\author{
Matthew Schumwinger
}
\keyword{datasets}

