% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIgeocoder.R
\docType{data}
\name{mai_04042016}
\alias{mai_04042016}
\title{mai_04042016 dataset}
\format{An object of class \code{data.frame} with 314920 rows and 15 columns.}
\usage{
mai_04042016
}
\description{
This contains the Master Address Index file downloaded and processed to a
data frame on April 4, 2016.
}
\examples{
data(mai_04042016)
head(mai_04042016)

}
\author{
Matthew Schumwinger
}
\keyword{datasets}

