% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mainroads}
\alias{mainroads}
\title{mainroads dataset}
\format{An object of class \code{SpatialLinesDataFrame} with 5370 rows and 10 columns.}
\usage{
mainroads
}
\description{
This contains the Open StreetMap main roads for metro Milwaukee shapefile
downloaded on Dec 3, 2012 and processed to a SpatialPolygonsDataFrame.
Mercator propjection.
}
\examples{
data(mainroads)
sp::plot(mainroads)

}
\author{
Matthew Schumwinger
}
\keyword{datasets}

