% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mkeoutline}
\alias{mkeoutline}
\title{mkeoutline dataset}
\format{An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 5 columns.}
\usage{
mkeoutline
}
\description{
This contains the corporate boundary of the City of Milwaukee shapefile
downloaded on Feb 26, 2015 and processed to a SpatialPolygonsDataFrame.
Projected using WGS84.
}
\examples{
data(mkeoutline)
sp::plot(mkeoutline)

}
\author{
Matthew Schumwinger
}
\keyword{datasets}

