% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mprop_12192016}
\alias{mprop_12192016}
\title{mprop_12192016 dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 161685 rows and 92 columns.}
\usage{
mprop_12192016
}
\description{
This contains the City of Milwaukee \href{http://itmdapps.milwaukee.gov/gis/mprop/Documentation/mprop.pdf}{MasterProperty Database (MPROP)}. The file was downloaded and processed to a data frame on December 19, 2016.
}
\examples{
head(mprop_12192016)
# join to parcels spdf
combo <- merge(parcels, mprop_12192016, by = TAXKEY, all.x = T)
# plot vacant land
plot(mkeoutline)
plot(subset(combo, LAND_USE_GP == "13"), add = T, border = "brown")

}
\author{
Matthew Schumwinger
}
\keyword{datasets}

