% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime_functions.R
\name{munge_wibrs}
\alias{munge_wibrs}
\title{munge_wibrs}
\usage{
munge_wibrs(df)
}
\arguments{
\item{df}{A data frame of WIBRS data.}
}
\value{
A SpatialPointsDataFrame in CRS NAD27.
}
\description{
munge_wibrs munges geo-processed WIBRS data, producing a
SpatialPointsDataFrame ready for analysis and visualization. Note: this
function does not exclude records with low or null geocode scores.
}
\examples{
data(crime.geo)
crimes <- munge_wibrs(crime.geo)
}

