% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw_wibrs_2005thru2015}
\alias{raw_wibrs_2005thru2015}
\title{raw_wibrs_2005thru2015 dataset}
\format{An object of class \code{data.frame} with 546464 rows and 10 columns.}
\usage{
raw_wibrs_2005thru2015
}
\description{
This contains all WIBRS crime incident records for the entire City of
Milwukee, from Jan 1, 2005 through Dec 31 2015. Records were downloaded from
the COMPASS website. 2015 data were retreived on 04/13/2016.
}
\examples{
data(raw_wibrs_2005thru2015)
head(raw_wibrs_2005thru2015)

}
\author{
Matthew Schumwinger
}
\keyword{datasets}

