% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{to_wgs84}
\alias{to_wgs84}
\title{to_wgs84}
\usage{
to_wgs84(x)
}
\value{
An sp class object.
}
\description{
to_wgs84 is a convenience function that transforms the CRS of an sp class
object to the WGS 1984 (EPSG:4326) coordinate system.
}
\examples{
\dontrun{
# plot 2015 homicides on web contextual layer
h2015 <- subset(crimes.munged, OFFENSE1 == "HOMICIDE" & year == "2015")
h2015 <- to_wgs84(h2015) # projection used for webmaps
m <- data.frame(slot(h2015, "coords"), slot(h2015, "data"))
names(m)[1:2] <- c("lon", "lat")
}

}

