% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wi_zipcodes}
\alias{wi_zipcodes}
\title{wi_zipcodes dataset}
\format{An object of class \code{SpatialPolygonsDataFrame} with 774 rows and 11 columns.}
\usage{
wi_zipcodes
}
\description{
This contains the 2010 ZIP Code shapes for Wisconsin, in NAD83.
}
\examples{
data(wi_zipcodes)
sp::plot(wi_zipcodes)

}
\author{
Matthew Schumwinger
}
\keyword{datasets}

