% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bids}
\alias{bids}
\title{bids dataset}
\format{An object of class \code{SpatialPolygonsDataFrame} with 35 rows and 2 columns.}
\usage{
bids
}
\description{
This contains the Milwaukee Business Improvement Distict (BID) shapefile
downloaded on Feb 26, 2015 and processed to a SpatialPolygonsDataFrame.
Projected using WGS84.
}
\examples{
data(bids)
sp::plot(bids)

}
\author{
Matthew Schumwinger
}
\keyword{datasets}

