% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_geocode.R
\name{full_geocode}
\alias{full_geocode}
\title{full_geocode}
\usage{
full_geocode(dataframe, field, MAI)
}
\arguments{
\item{dataframe}{A dataframe with City of Milwaukee records to geocode.}

\item{field}{(NOT IMPLEMENTED)Character string of single column name containing street address
information to geocode. (Note: you may need to concatenate prior to using
this function).}

\item{MAI}{A data frame containing the Master Address Index, already
geoprocessed against parcel centroids.}
}
\value{
A data frame.
}
\description{
full_geocode is a function that geocode addresses in Milwaukee,
Wisconsin. Draft version geocodes crime data, first by matching against
geoprocessed MAI, thne using low_level funciton to access City geocode API.
}
\examples{
\dontrun{
data("raw_wibrs_2005thru2015")
data("geo_mai")
geo_crime <- full_geocode(raw_wibrs_2005thru2015, "LOCATION", geo_mai)
head(geo_crime)
}

}

