% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIgeocoder.R
\name{geocode}
\alias{geocode}
\title{geocode}
\usage{
geocode(batch, fields)
}
\arguments{
\item{batch}{Data frame to be geocoded.}

\item{fields}{Character vector of batch fields containing street address.}
}
\value{
A data frame of the original batch data frame, with new geocoding fields added.
}
\description{
geocode is a funtion that uses City of Milwaukee Public geocode API to
geocode addresses. (Specifically, the \href{http://maps.milwaukee.gov/ArcGIS/rest/services/
geocode/MAIthenDIME_geocode/GeocodeServer}{MAI-then-DIME} method.)
}
\examples{
\dontrun{
raw_geo <- geocode(head(raw), "LOCATION")
raw_geo
}

}

