% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_bids}
\alias{get_bids}
\title{get_bids}
\usage{
get_bids()
}
\value{
A large SpatialPolygonsDataFrame.
}
\description{
get_bids is a function that retreives the City of Milwaukee's Business
Improvement District (BID) shapefile
\href{http://city.milwaukee.gov/DownloadMapData3497.htm}{from the
city's download map data page}.
}
\examples{
\dontrun{
bids <- get_bids()
sp::plot(bids)
head(bids@data)
}

}

