% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mprop_12192016}
\alias{mprop_12192016}
\title{mprop_12192016 dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 161685 rows and 92 columns.}
\usage{
mprop_12192016
}
\description{
This contains the City of Milwaukee \href{http://itmdapps.milwaukee.gov/gis/mprop/Documentation/mprop.pdf}
{MasterProperty Database (MPROP)}. The file was downloaded and processed to a
data frame on December 19, 2016.
}
\examples{
data(mprop_12192016)
head(mprop_12192016)

}
\author{
Matthew Schumwinger
}
\keyword{datasets}

