% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{to_nad27}
\alias{to_nad27}
\title{to_nad27}
\usage{
to_nad27(x)
}
\value{
An sp class object.
}
\description{
to_nad27 is a convenience function that transforms the CRS of an sp class
object to the State Plane South NAD 27 (ESPG 32054) datum.
}
\examples{
\dontrun{
# plot ZIP codes in and around Milwuakee
wi_zipcodes <- to_nad27(wi_zipcodes)
plot(mkeoutline, border = "red", lwd = 5)
plot(wi_zipcodes, add = T)
}

}

