% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{to_wgs84}
\alias{to_wgs84}
\title{to_wgs84}
\usage{
to_wgs84(x)
}
\value{
An sp class object.
}
\description{
to_wgs84 is a convenience function that transforms the CRS of an sp class
object to the WGS 1984 (EPSG:4326) coordinate system.
}
\examples{
\dontrun{
hoods <- to_wgs84(hoods)
plot(hoods)
plot(bids, col = "salmon", add = T)
}

}

